Added module for translating a metaxa description to a rascal syntax description.
At this point there is no working analysis module, i.e. a module to
check if a metaxa description makes sense with regards to typing
etc. The plan right now is to assume that any incoming description is on
the right form, and generate a rascal file based on that. We will simply
have to make sure that our example descriptions are correct without
using any automated semantic analysis.