Added an AST module. This module defines the data types which can be used together with the `implode` function to make an AST out of a parse tree.
Note that the AST is not complete; it can not be used to translate an
arbitrary parse tree to the corresponding AST. I have only implemented
the data types that I need at this point.

The motivation for adding an AST module was because dealing with the
parse tree was annoying when it came to star-productions. It seemed like
I had to implement custom helper functions in order to extract any
star-production to a list of those productions; if I just used the
star-production as-is, it didn't seem like I could use it as a
collection (like list, set, etc.).