Added a new example, which is slightly more complex than the helloworld.mtx; a grammar consisting of three alternative constant strings.
This description can successfully be translated to rascal right now.

Though there could perhaps be another way for such a grammar to be
defined, which we could implement. Something like this:

construct Alts {

syntax { option_1 | option_2 | option_3 }

}

But this doesn't seem to be possible for such constant strings, given the current grammar (only
SortExpr productions can be listed as alternatives like this).